% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_summary.R
\name{general_summary}
\alias{general_summary}
\title{General Summary Function}
\usage{
general_summary(
  data,
  variable,
  group_vars = NULL,
  weights_var = NULL,
  multiple_choice = FALSE,
  variable_type = "binary"
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{variable}{The name of the variable to summarize.}

\item{group_vars}{An optional vector of variable names by which to group the summary.}

\item{weights_var}{An optional variable name containing weights for weighted summaries.}

\item{multiple_choice}{A logical indicating whether the variable is a multiple choice variable. Default is FALSE.}

\item{variable_type}{The type of the variable: "binary", "continuous", or other. Default is "binary".}
}
\value{
A data frame containing the summary.
}
\description{
This function provides a summary of a given variable within a data frame, optionally grouped by one or more other variables.
It supports binary, continuous, and multiple choice variables.
}
\examples{
\dontrun{
general_summary(data = my_data, variable = "my_var", group_vars = c("group1", "group2"))
}

}
